--- pkg/i18n/spanish.go
@@ -1,129 +1,0 @@
-package i18n
-
-func spanishSet() TranslationSet {
-	return TranslationSet{
-		PruningStatus:              "limpiando",
-		RemovingStatus:             "eliminando",
-		RestartingStatus:           "reiniciando",
-		StartingStatus:             "iniciando",
-		StoppingStatus:             "terminando",
-		UppingServiceStatus:        "levantando servicio",
-		UppingProjectStatus:        "levantando proyecto",
-		DowningStatus:              "dando de baja",
-		PausingStatus:              "pausando",
-		RunningCustomCommandStatus: "ejecutando comando personalizado",
-		RunningBulkCommandStatus:   "ejecutando comando masivo",
-
-		ErrorOccurred:                 "¡Hubo un error! Por favor crea un issue en https://github.com/jesseduffield/lazydocker/issues",
-		ConnectionFailed:              "Falló la conexión con el docker client. Quizá necesitas reiniciar tu docker client",
-		UnattachableContainerError:    "Container does not support attaching. You must either run the service with the '-it' flag or use `stdin_open: true, tty: true` in the docker-compose.yml file",
-		WaitingForContainerInfo:       "No podemos proceder hasta que docker nos de más información sobre el contenedor. Inténtalo otra vez en unos segundos.",
-		CannotAccessDockerSocketError: "No es posible acceder al docker socket en: unix:///var/run/docker.sock\nEjecuta lazydocker como root o lee https://docs.docker.com/install/linux/linux-postinstall/",
-		CannotKillChildError:          "Esperamos tres segundos a que el proceso hijo se detenga. Debe de haber un proceso huérfano que continua activo en tu sistema.",
-
-		Donate:  "Donar",
-		Confirm: "Confirmar",
-
-		Return:                      "regresar",
-		FocusMain:                   "enfocar panel principal",
-		LcFilter:                    "filtrar lista",
-		Navigate:                    "navegar",
-		Execute:                     "ejecutar",
-		Close:                       "cerrar",
-		Quit:                        "salir",
-		Menu:                        "menú",
-		MenuTitle:                   "Menú",
-		OpenConfig:                  "abrir configuración de lazydocker",
-		EditConfig:                  "editar configuración de lazydocker",
-		Cancel:                      "cancelar",
-		Remove:                      "borrar",
-		HideStopped:                 "esconder/mostrar contenedores parados",
-		ForceRemove:                 "borrar(forzado)",
-		RemoveWithVolumes:           "borrar con volúmenes",
-		RemoveService:               "borrar contenedores",
-		UpService:                   "levantar servicio",
-		Stop:                        "parar",
-		Pause:                       "pausa",
-		Restart:                     "reiniciar",
-		Down:                        "bajar proyecto",
-		DownWithVolumes:             "bajar proyecto con volúmenes",
-		Start:                       "iniciar",
-		Rebuild:                     "recompilar",
-		Recreate:                    "recrear",
-		PreviousContext:             "anterior pestaña",
-		NextContext:                 "siguiente pestaña",
-		ViewLogs:                    "ver logs",
-		UpProject:                   "levantar proyecto",
-		DownProject:                 "dar de baja el proyecto",
-		RemoveImage:                 "limpiar imagen",
-		RemoveVolume:                "limpiar volúmen",
-		RemoveNetwork:               "limpiar red",
-		RemoveWithoutPrune:          "limpiar sin borrar padres sin etiqueta",
-		RemoveWithoutPruneWithForce: "limpiar (forzado) sin borrar padres sin etiqueta",
-		RemoveWithForce:             "limpiar (forzado)",
-		PruneContainers:             "limpiar contenedores finalizados",
-		PruneVolumes:                "limpiar volúmenes sin usar",
-		PruneNetworks:               "limpiar redes sin usar",
-		PruneImages:                 "limpiar imágenes sin usar",
-		StopAllContainers:           "detener todos los contenedores",
-		RemoveAllContainers:         "borrar todos los contenedores (forzado)",
-		ViewRestartOptions:          "ver opciones de reinicio",
-		ExecShell:                   "ejecutar shell",
-		RunCustomCommand:            "ejecutar comando personalizado",
-		ViewBulkCommands:            "ver comandos masivos",
-		FilterList:                  "filtar list",
-		OpenInBrowser:               "abrir en navegador (first port is http)",
-		SortContainersByState:       "ordenar contenedores por estado",
-
-		GlobalTitle:               "Global",
-		MainTitle:                 "Inicio",
-		ProjectTitle:              "Proyecto",
-		ServicesTitle:             "Servicios",
-		ContainersTitle:           "Contenedores",
-		StandaloneContainersTitle: "Contenedores independientes",
-		ImagesTitle:               "Imágenes",
-		VolumesTitle:              "Volúmenes",
-		NetworksTitle:             "Redes",
-		CustomCommandTitle:        "Comando personalizado:",
-		BulkCommandTitle:          "Comando masivo:",
-		ErrorTitle:                "Error",
-		LogsTitle:                 "Logs",
-		ConfigTitle:               "Configuración",
-		EnvTitle:                  "Variables de entorno",
-		DockerComposeConfigTitle:  "Docker-Compose Config",
-		TopTitle:                  "Top",
-		StatsTitle:                "Estadísticas",
-		CreditsTitle:              "Acerca",
-		ContainerConfigTitle:      "Configuración",
-		ContainerEnvTitle:         "Variables de entorno",
-		NothingToDisplay:          "Nada que mostrar",
-		NoContainerForService:     "No hay logs que mostrar; el servicio no está asociado con un contenedor",
-		CannotDisplayEnvVariables: "Algo salió mal mientras se mostraban las variables de entorno",
-
-		NoContainers: "Sin contenedores",
-		NoContainer:  "Sin contenedor",
-		NoImages:     "Sin imágenes",
-		NoVolumes:    "Sin volúmenes",
-		NoNetworks:   "Sin redes",
-		NoServices:   "Sin servicios",
-
-		ConfirmQuit:                "¿Realmente quieres salir?",
-		ConfirmUpProject:           "¿Realmente quieres levantar tu proyecto docker compose?",
-		MustForceToRemoveContainer: "No puedes borrar un contenedor en ejecución a menos de que lo fuerces, ¿quieres hacerlo?",
-		NotEnoughSpace:             "No hay suficiente espacio para renderizar los paneles",
-		ConfirmPruneImages:         "¿Realmente quieres limpiar todas tus imágenes?",
-		ConfirmPruneContainers:     "¿Realmente quieres limpiar todos los contenedores finalizados?",
-		ConfirmStopContainers:      "¿Realmente quieres detener todos los contenedores?",
-		ConfirmRemoveContainers:    "¿Realmente quieres borrar todos los contenedores?",
-		ConfirmPruneVolumes:        "¿Realmente quieres limpiar todos los vólumenes sin usar?",
-		ConfirmPruneNetworks:       "¿Realmente quieres limpiar todas las redes sin usar?",
-		StopService:                "¿Realmente quieres detener los contenedores de este servicio?",
-		StopContainer:              "¿Realmente quieres detener este contenedor?",
-		PressEnterToReturn:         "Presionar [enter] para volver a lazydocker (este mensaje puede ser desactivado en tu configuración poniendo `gui.returnImmediately: true`)",
-
-		No:  "no",
-		Yes: "sí",
-
-		FilterPrompt: "filtrar",
-	}
-}
