+++ pkg/i18n/translations/az.json
@@ -1,0 +1,134 @@
+{
+  "code": "az",
+  "name": "Azərbaycanca",
+  "translations": {
+    "PruningStatus": "təmizlənir",
+    "RemovingStatus": "silinir",
+    "RestartingStatus": "yenidən başladılır",
+    "StartingStatus": "başladılır",
+    "StoppingStatus": "dayandırılır",
+    "UppingServiceStatus": "servis qaldırılır",
+    "UppingProjectStatus": "layihə qaldırılır",
+    "DowningStatus": "söndürülür",
+    "PausingStatus": "dayandırılır",
+    "RunningCustomCommandStatus": "xüsusi əmr icra edilir",
+    "RunningBulkCommandStatus": "toplu əmr icra edilir",
+    "NoViewMachingNewLineFocusedSwitchStatement": "newLineFocused switch statement ilə uyğun görünüş yoxdur",
+    "ErrorOccurred": "Xəta baş verdi! Zəhmət olmasa https://github.com/jesseduffield/lazydocker/issues ünvanında məsələ yaradın",
+    "ConnectionFailed": "Docker müştərisi ilə əlaqə uğursuz oldu. Docker müştərisini yenidən başlatmalı ola bilərsiniz",
+    "UnattachableContainerError": "Konteyner qoşulmağı dəstəkləmir. Xidməti '-it' bayrağı ilə işə salmalı və ya docker-compose.yml faylında `stdin_open: true, tty: true` istifadə etməlisiniz",
+    "WaitingForContainerInfo": "Docker bizə konteyner haqqında daha çox məlumat verənə qədər davam edə bilmərik. Zəhmət olmasa bir neçə dəqiqədən sonra yenidən cəhd edin.",
+    "CannotAttachStoppedContainerError": "Dayandırılmış konteynerə qoşula bilməzsiniz, əvvəlcə onu başlatmalısınız ('r' düyməsi ilə bunu edə bilərsiniz) (bəli, bunu sizin üçün avtomatik etmək üçün çox tənbələm) (xəta mesajı formasında sizinlə bir-birə ünsiyyət qurmaq çox maraqlıdır)",
+    "CannotAccessDockerSocketError": "Docker socket-ə daxil olmaq mümkün deyil: unix:///var/run/docker.sock\nLazydocker-i root olaraq işə salın və ya https://docs.docker.com/install/linux/linux-postinstall/ oxuyun",
+    "CannotKillChildError": "Uşaq prosesinin dayanması üçün üç saniyə gözlənildi. Sisteminizdə davam edən yetim proses ola bilər.",
+    "Donate": "İanə et",
+    "Confirm": "Təsdiq et",
+    "Return": "geri qayıt",
+    "FocusMain": "əsas panelə fokuslan",
+    "LcFilter": "siyahını filtrle",
+    "Navigate": "naviqasiya et",
+    "Execute": "icra et",
+    "Close": "bağla",
+    "Quit": "çıx",
+    "Menu": "menyu",
+    "MenuTitle": "Menyu",
+    "Scroll": "sürüşdür",
+    "OpenConfig": "lazydocker konfiqurasiyasını aç",
+    "EditConfig": "lazydocker konfiqurasiyasını redaktə et",
+    "Cancel": "ləğv et",
+    "Remove": "sil",
+    "HideStopped": "dayandırılmış konteynerləri gizlət/göstər",
+    "ForceRemove": "məcburi sil",
+    "RemoveWithVolumes": "həcmlərlə birlikdə sil",
+    "RemoveService": "konteynerləri sil",
+    "UpService": "xidməti qaldır",
+    "Stop": "dayandır",
+    "Pause": "fasilə",
+    "Restart": "yenidən başlat",
+    "Down": "layihəni söndür",
+    "DownWithVolumes": "layihəni həcmlərlə birlikdə söndür",
+    "Start": "başlat",
+    "Rebuild": "yenidən qur",
+    "Recreate": "yenidən yarat",
+    "PreviousContext": "əvvəlki tab",
+    "NextContext": "növbəti tab",
+    "Attach": "qoşul",
+    "ViewLogs": "logları göstər",
+    "UpProject": "layihəni qaldır",
+    "DownProject": "layihəni söndür",
+    "RemoveImage": "təsviri sil",
+    "RemoveVolume": "həcmi sil",
+    "RemoveNetwork": "şəbəkəni sil",
+    "RemoveWithoutPrune": "etiketlənməmiş valideynləri silmədən sil",
+    "RemoveWithoutPruneWithForce": "sil (məcburi) etiketlənməmiş valideynləri silmədən",
+    "RemoveWithForce": "sil (məcburi)",
+    "PruneContainers": "çıxılmış konteynerləri təmizlə",
+    "PruneVolumes": "istifadə olunmayan həcmləri təmizlə",
+    "PruneNetworks": "istifadə olunmayan şəbəkələri təmizlə",
+    "PruneImages": "istifadə olunmayan təsvirləri təmizlə",
+    "StopAllContainers": "bütün konteynerləri dayandır",
+    "RemoveAllContainers": "bütün konteynerləri sil (məcburi)",
+    "ViewRestartOptions": "yenidən başlatma variantlarını göstər",
+    "ExecShell": "shell icra et",
+    "RunCustomCommand": "əvvəlcədən təyin edilmiş xüsusi əmr icra et",
+    "ViewBulkCommands": "toplu əmrləri göstər",
+    "FilterList": "siyahını filtrle",
+    "OpenInBrowser": "brauzerdə aç (ilk port http-dir)",
+    "SortContainersByState": "konteynerləri vəziyyətə görə sırala",
+    "GlobalTitle": "Qlobal",
+    "MainTitle": "Əsas",
+    "ProjectTitle": "Layihə",
+    "ServicesTitle": "Xidmətlər",
+    "ContainersTitle": "Konteynerlər",
+    "StandaloneContainersTitle": "Müstəqil Konteynerlər",
+    "ImagesTitle": "Təsvirlər",
+    "VolumesTitle": "Həcmlər",
+    "NetworksTitle": "Şəbəkələr",
+    "CustomCommandTitle": "Xüsusi Əmr:",
+    "BulkCommandTitle": "Toplu Əmr:",
+    "ErrorTitle": "Xəta",
+    "LogsTitle": "Loglar",
+    "ConfigTitle": "Konfiqurasiya",
+    "EnvTitle": "Mühit",
+    "DockerComposeConfigTitle": "Docker-Compose Konfiqurasiyası",
+    "TopTitle": "Ən yuxarı",
+    "StatsTitle": "Statistika",
+    "CreditsTitle": "Haqqında",
+    "ContainerConfigTitle": "Konteyner Konfiqurasiyası",
+    "ContainerEnvTitle": "Konteyner Mühiti",
+    "NothingToDisplay": "Göstəriləcək heç nə yoxdur",
+    "NoContainerForService": "Göstəriləcək log yoxdur; xidmət konteyner ilə əlaqələndirilməyib",
+    "CannotDisplayEnvVariables": "Mühit dəyişənləri göstərilərkən xəta baş verdi",
+    "NoContainers": "Konteyner yoxdur",
+    "NoContainer": "Konteyner yoxdur",
+    "NoImages": "Təsvir yoxdur",
+    "NoVolumes": "Həcm yoxdur",
+    "NoNetworks": "Şəbəkə yoxdur",
+    "NoServices": "Xidmət yoxdur",
+    "ConfirmQuit": "Çıxmaq istədiyinizə əminsiniz?",
+    "ConfirmUpProject": "Docker compose layihənizi 'qaldırmaq' istədiyinizə əminsiniz?",
+    "MustForceToRemoveContainer": "İşləyən konteyneri məcbur etmədən silə bilməzsiniz. Məcbur etmək istəyirsiniz?",
+    "NotEnoughSpace": "Panelləri göstərmək üçün kifayət qədər yer yoxdur",
+    "ConfirmPruneImages": "Bütün istifadə olunmayan təsvirləri təmizləmək istədiyinizə əminsiniz?",
+    "ConfirmPruneContainers": "Bütün dayandırılmış konteynerləri təmizləmək istədiyinizə əminsiniz?",
+    "ConfirmStopContainers": "Bütün konteynerləri dayandırmaq istədiyinizə əminsiniz?",
+    "ConfirmRemoveContainers": "Bütün konteynerləri silmək istədiyinizə əminsiniz?",
+    "ConfirmPruneVolumes": "Bütün istifadə olunmayan həcmləri təmizləmək istədiyinizə əminsiniz?",
+    "ConfirmPruneNetworks": "Bütün istifadə olunmayan şəbəkələri təmizləmək istədiyinizə əminsiniz?",
+    "StopService": "Bu xidmətin konteynerlərini dayandırmaq istədiyinizə əminsiniz?",
+    "StopContainer": "Bu konteyneri dayandırmaq istədiyinizə əminsiniz?",
+    "PressEnterToReturn": "Lazydocker-ə qayıtmaq üçün enter düyməsini basın (bu bildiriş konfiqurasiyada `gui.returnImmediately: true` təyin edərək deaktiv edilə bilər)",
+    "DetachFromContainerShortCut": "Defolt olaraq, konteynerdən ayrılmaq üçün ctrl-p, sonra ctrl-q basın",
+    "No": "xeyr",
+    "Yes": "bəli",
+    "LcNextScreenMode": "növbəti ekran rejimi (normal/yarım/tam ekran)",
+    "LcPrevScreenMode": "əvvəlki ekran rejimi",
+    "FilterPrompt": "filtr",
+    "FocusProjects": "layihələr panelinə fokuslan",
+    "FocusServices": "xidmətlər panelinə fokuslan",
+    "FocusContainers": "konteynerlər panelinə fokuslan",
+    "FocusImages": "təsvirlər panelinə fokuslan",
+    "FocusVolumes": "həcmlər panelinə fokuslan",
+    "FocusNetworks": "şəbəkələr panelinə fokuslan"
+  }
+}
