+++ pkg/i18n/translations/fr.json
@@ -1,0 +1,111 @@
+{
+  "code": "fr",
+  "name": "Français",
+  "translations": {
+    "PruningStatus": "destruction",
+    "RemovingStatus": "suppression",
+    "RestartingStatus": "redémarrage",
+    "StartingStatus": "démarrage",
+    "StoppingStatus": "arrêt",
+    "PausingStatus": "mise en pause",
+    "RunningCustomCommandStatus": "exécution de la commande personalisée",
+    "RunningBulkCommandStatus": "exécution de la commande groupée",
+    "NoViewMachingNewLineFocusedSwitchStatement": "Aucune vue correspondant au switch newLineFocused",
+    "ErrorOccurred": "Une erreur s'est produite ! Veuillez créer un rapport d'erreur sur https://github.com/jesseduffield/lazydocker/issues",
+    "ConnectionFailed": "Erreur lors de la connexion au client Docker. Essayez de redémarrer votre client Docker",
+    "UnattachableContainerError": "Le conteneur ne peut pas être attaché. Vous devez exécuter le service avec le drapeau 'it' ou bien utiliser `stdin_open: true, tty: true` dans votre fichier docker-compose.yml",
+    "WaitingForContainerInfo": "Le processus ne peut pas continuer avant que Docker ne fournisse plus d'informations. Veuillez réessayer dans quelques instants.",
+    "CannotAttachStoppedContainerError": "Vous ne pouvez pas vous attacher à un conteneur arrêté, vous devez le démarrer en amont (ce que vous pouvez faire avec la touche 'r') (oui, je suis trop paresseux pour le faire automatiquement pour vous) (plutôt cool que je puisse communiquer en tête-à-tête avec vous au travers d'un message d'erreur, cependant)",
+    "CannotAccessDockerSocketError": "Impossible d'accéder au socket Docker à : unix:///var/run/docker.sock\nLancez lazydocker en tant que root ou alors lisez https://docs.docker.com/install/linux/linux-postinstall/",
+    "CannotKillChildError": "Trois secondes se sont écoulées depuis la demande d'arrêt des processus enfants. Il se peut qu'un processus orphelin continue à tourner sur votre système.",
+    "Donate": "Donner",
+    "Confirm": "Confirmer",
+    "Return": "retour",
+    "FocusMain": "focus panneau principal",
+    "Navigate": "naviguer",
+    "Execute": "exécuter",
+    "Close": "fermer",
+    "Menu": "menu",
+    "MenuTitle": "Menu",
+    "Scroll": "faire défiler",
+    "OpenConfig": "ouvrir la configuration lazydocker",
+    "EditConfig": "modifier la configuration lazydocker",
+    "Cancel": "annuler",
+    "Remove": "supprimer",
+    "HideStopped": "cacher/montrer les conteneurs arrêtés",
+    "ForceRemove": "forcer la suppression",
+    "RemoveWithVolumes": "supprimer avec les volumes",
+    "RemoveService": "supprimer les conteneurs",
+    "Stop": "arrêter",
+    "Pause": "pause",
+    "Restart": "redémarrer",
+    "Start": "démarrer",
+    "Rebuild": "reconstruire",
+    "Recreate": "recréer",
+    "PreviousContext": "onglet précédent",
+    "NextContext": "onglet suivant",
+    "Attach": "attacher",
+    "ViewLogs": "voir les enregistrements",
+    "RemoveImage": "supprimer l'image",
+    "RemoveVolume": "supprimer le volume",
+    "RemoveNetwork": "supprimer le réseau",
+    "RemoveWithoutPrune": "supprimer sans effacer les parents non étiquetés",
+    "RemoveWithoutPruneWithForce": "supprimer (forcer) sans effacer les parents non étiquetés",
+    "RemoveWithForce": "supprimer (forcer)",
+    "PruneContainers": "détruire les conteneurs arrêtés",
+    "PruneVolumes": "détruire les volumes non utilisés",
+    "PruneNetworks": "détruire les réseaux non utilisés",
+    "PruneImages": "détruire les images non utilisées",
+    "StopAllContainers": "arrêter tous les conteneurs",
+    "RemoveAllContainers": "supprimer tous les conteneurs (forcer)",
+    "ViewRestartOptions": "voir les options de redémarrage",
+    "ExecShell": "exécuter le shell",
+    "RunCustomCommand": "exécuter une commande prédéfinie",
+    "ViewBulkCommands": "voir les commandes groupées",
+    "OpenInBrowser": "ouvrir dans le navigateur (le premier port est http)",
+    "SortContainersByState": "ordonner les conteneurs par état",
+    "GlobalTitle": "Global",
+    "MainTitle": "Principal",
+    "ProjectTitle": "Projet",
+    "ServicesTitle": "Services",
+    "ContainersTitle": "Conteneurs",
+    "StandaloneContainersTitle": "Conteneurs autonomes",
+    "ImagesTitle": "Images",
+    "VolumesTitle": "Volumes",
+    "NetworksTitle": "Réseaux",
+    "CustomCommandTitle": "Commande personnalisée :",
+    "BulkCommandTitle": "Commande groupée :",
+    "ErrorTitle": "Erreur",
+    "LogsTitle": "Journaux",
+    "ConfigTitle": "Config",
+    "EnvTitle": "Env",
+    "DockerComposeConfigTitle": "Config Docker-Compose",
+    "TopTitle": "Top",
+    "StatsTitle": "Statistiques",
+    "CreditsTitle": "À propos",
+    "ContainerConfigTitle": "Config Conteneur",
+    "ContainerEnvTitle": "Env Conteneur",
+    "NothingToDisplay": "Rien à afficher",
+    "CannotDisplayEnvVariables": "Quelque chose a échoué lors de l'affichage des variables d'environnement",
+    "NoContainers": "Aucun conteneur",
+    "NoContainer": "Aucun conteneur",
+    "NoImages": "Aucune image",
+    "NoVolumes": "Aucun volume",
+    "NoNetworks": "Aucun réseau",
+    "ConfirmQuit": "Êtes-vous certain de vouloir quitter ?",
+    "MustForceToRemoveContainer": "Vous ne pouvez pas supprimer un conteneur qui tourne sans le forcer. Voulez-vous le forcer ?",
+    "NotEnoughSpace": "Manque d'espace pour afficher les différent panneaux",
+    "ConfirmPruneImages": "Êtes-vous certain de vouloir détruire toutes les images non utilisées ?",
+    "ConfirmPruneContainers": "Êtes-vous certain de vouloir détruire tous les conteneurs arrêtés ?",
+    "ConfirmStopContainers": "Êtes-vous certain de vouloir arrêter tous les conteneurs ?",
+    "ConfirmRemoveContainers": "Êtes-vous certain de vouloir supprimer tous les conteneurs ?",
+    "ConfirmPruneVolumes": "Êtes-vous certain de vouloir détruire tous les volumes non utilisés ?",
+    "ConfirmPruneNetworks": "Êtes-vous certain de vouloir détruire tous les réseaux non utilisés ?",
+    "StopService": "Êtes-vous certain de vouloir arrêter le conteneur de ce service ?",
+    "StopContainer": "Êtes-vous certain de vouloir arrêter ce conteneur ?",
+    "PressEnterToReturn": "Appuyez sur Entrée pour revenir à lazydocker (ce message peut être désactivé dans vos configurations en appliquant `gui.returnImmediately: true`)",
+    "DetachFromContainerShortCut": "Par défaut, pour se détacher du conteneur appuyez sur CTRL-P puis CTRL-Q",
+    "No": "non",
+    "Yes": "oui"
+  }
+}
