+++ pkg/i18n/translations/pt.json
@@ -1,0 +1,128 @@
+{
+  "code": "pt",
+  "name": "Português",
+  "translations": {
+    "PruningStatus": "destruindo",
+    "RemovingStatus": "removendo",
+    "RestartingStatus": "reiniciando",
+    "StartingStatus": "iniciando",
+    "StoppingStatus": "parando",
+    "UppingServiceStatus": "subindo serviço",
+    "UppingProjectStatus": "subindo projeto",
+    "DowningStatus": "derrubando",
+    "PausingStatus": "pausando",
+    "RunningCustomCommandStatus": "executando comando personalizado",
+    "RunningBulkCommandStatus": "executando comando em massa",
+    "NoViewMachingNewLineFocusedSwitchStatement": "No view matching newLineFocused switch statement",
+    "ErrorOccurred": "Um erro ocorreu! Por favor, crie uma issue em https://github.com/jesseduffield/lazydocker/issues",
+    "ConnectionFailed": "Falha na conexão com o cliente Docker. Você pode precisar reiniciar o seu cliente Docker",
+    "UnattachableContainerError": "O contêiner não suporta anexação. Você deve executar o serviço com a flag '-it' ou usar `stdin_open: true, tty: true` no arquivo docker-compose.yml",
+    "WaitingForContainerInfo": "Não é possível prosseguir até que o Docker forneça mais informações sobre o contêiner. Por favor, tente novamente em alguns momentos.",
+    "CannotAttachStoppedContainerError": "Você não pode anexar a um contêiner parado, você precisa iniciá-lo primeiro (o que você pode fazer com a tecla 'r') (sim, sou preguiçoso demais para fazer isso automaticamente para você) (aliás, bem legal que eu posso me comunicar diretamente com você na forma de uma mensagem de erro)",
+    "CannotAccessDockerSocketError": "Não é possível acessar o sôquete docker em: unix:///var/run/docker.sock\nExecute o lazydocker como root ou leia https://docs.docker.com/install/linux/linux-postinstall/",
+    "CannotKillChildError": "Três segundos foram esperarados para que os processos filhos parassem. Pode haver um processo órfão que continua em execução em seu sistema.",
+    "Donate": "Doar",
+    "Confirm": "Confirmar",
+    "Return": "retornar",
+    "FocusMain": "focar no painel principal",
+    "LcFilter": "filtrar lista",
+    "Navigate": "navegar",
+    "Execute": "executar",
+    "Close": "fechar",
+    "Quit": "sair",
+    "Menu": "menu",
+    "MenuTitle": "Menu",
+    "Scroll": "rolar",
+    "OpenConfig": "abrir configuração do lazydocker",
+    "EditConfig": "editar configuração do lazydocker",
+    "Cancel": "cancelar",
+    "Remove": "remover",
+    "HideStopped": "ocultar/mostrar contêineres parados",
+    "ForceRemove": "forçar remoção",
+    "RemoveWithVolumes": "remover com volumes",
+    "RemoveService": "remover contêineres",
+    "UpService": "subir serviço",
+    "Stop": "parar",
+    "Pause": "pausar",
+    "Restart": "reiniciar",
+    "Down": "derrubar projeto",
+    "DownWithVolumes": "derrubar projetos com volumes",
+    "Start": "iniciar",
+    "Rebuild": "reconstuir",
+    "Recreate": "recriar",
+    "PreviousContext": "aba anterior",
+    "NextContext": "próxima aba",
+    "Attach": "anexar",
+    "ViewLogs": "ver logs",
+    "UpProject": "subir projeto",
+    "DownProject": "derrubar projeto",
+    "RemoveImage": "remover imagem",
+    "RemoveVolume": "remover volume",
+    "RemoveNetwork": "remover rede",
+    "RemoveWithoutPrune": "remover sem deletar pais não etiquetados",
+    "RemoveWithoutPruneWithForce": "remover (forçado) sem deletar pais não etiquetados",
+    "RemoveWithForce": "remover (forçado)",
+    "PruneContainers": "destruir contêineres encerrados",
+    "PruneVolumes": "destruir volumes não utilizados",
+    "PruneNetworks": "destruir redes não utilizadas",
+    "PruneImages": "destruir imagens não utilizadas",
+    "StopAllContainers": "parar todos os contêineres",
+    "RemoveAllContainers": "remover todos os contêineres (forçado)",
+    "ViewRestartOptions": "ver opções de reinício",
+    "ExecShell": "executar shell",
+    "RunCustomCommand": "executar comando personalizado predefinido",
+    "ViewBulkCommands": "ver comandos em massa",
+    "FilterList": "filtrar lista",
+    "OpenInBrowser": "abrir no navegador (primeira porta é http)",
+    "SortContainersByState": "ordenar contêineres por estado",
+    "GlobalTitle": "Global",
+    "MainTitle": "Principal",
+    "ProjectTitle": "Projeto",
+    "ServicesTitle": "Serviços",
+    "ContainersTitle": "Contêineres",
+    "StandaloneContainersTitle": "Contêineres Avulsos",
+    "ImagesTitle": "Imagens",
+    "VolumesTitle": "Volumes",
+    "NetworksTitle": "Redes",
+    "CustomCommandTitle": "Comando Personalizado:",
+    "BulkCommandTitle": "Comando em Massa:",
+    "ErrorTitle": "Erro",
+    "LogsTitle": "Registros",
+    "ConfigTitle": "Config",
+    "EnvTitle": "Env",
+    "DockerComposeConfigTitle": "Docker-Compose Config",
+    "TopTitle": "Topo",
+    "StatsTitle": "Estatísticas",
+    "CreditsTitle": "Sobre",
+    "ContainerConfigTitle": "Configuração do Contêiner",
+    "ContainerEnvTitle": "Contêiner Env",
+    "NothingToDisplay": "Nada a exibir",
+    "NoContainerForService": "Nenhum log para exibir; o serviço não está associado a nenhum contêiner",
+    "CannotDisplayEnvVariables": "Algo deu errado ao exibir as variáveis de ambiente",
+    "NoContainers": "Sem contêineres",
+    "NoContainer": "Sem contêiner",
+    "NoImages": "Sem imagens",
+    "NoVolumes": "Sem volumes",
+    "NoNetworks": "Sem redes",
+    "NoServices": "Sem serviços",
+    "ConfirmQuit": "Tem certeza que deseja sair?",
+    "ConfirmUpProject": "Tem certeza que deseja 'iniciar' seu projeto docker compose?",
+    "MustForceToRemoveContainer": "Você não pode remover um contêiner em execução a menos que o force. Deseja forçar?",
+    "NotEnoughSpace": "Sem espaço suficiente para renderizar os painéis",
+    "ConfirmPruneImages": "Tem certeza que deseja eliminar todas as imagens não utilizadas?",
+    "ConfirmPruneContainers": "Tem certeza que deseja destruir todos os contêineres parados?",
+    "ConfirmStopContainers": "Tem certeza que deseja parar todos os contêineres?",
+    "ConfirmRemoveContainers": "Tem certeza que deseja remover todos os contêineres?",
+    "ConfirmPruneVolumes": "Tem certeza que deseja destruir todos os volumes não utilizados?",
+    "ConfirmPruneNetworks": "Tem certeza que deseja destruir todas as redes não utilizadas?",
+    "StopService": "Tem certeza que deseja parar os contêineres deste serviço?",
+    "StopContainer": "Tem certeza que deseja parar este contêiner?",
+    "PressEnterToReturn": "Pressione enter para retornar ao lazydocker (este prompt pode ser desativado em sua configuração definindo `gui.returnImmediately: true`)",
+    "DetachFromContainerShortCut": "Por padrão, para desanexar do contêiner, pressione ctrl-p e depois ctrl-q",
+    "No": "não",
+    "Yes": "sim",
+    "LcNextScreenMode": "modo de tela seguinte (normal/meia/tela cheia)",
+    "LcPrevScreenMode": "modo de tela anterior",
+    "FilterPrompt": "filtro"
+  }
+}
